backend/requirements1.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.2.10
boto==2.49.0
boto3==1.16.25
botocore==1.19.25
certifi==2020.6.20
chardet==3.0.4
comtypes==1.1.7
coverage==5.3
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django
django-allauth==0.42.0
django-cors-headers==3.7.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-multiforloop==0.2.1
django-restframework==0.0.1
django-secure==1.0.2
django-storages==1.10.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
timedelta==2020.12.3
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jmespath==0.10.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
mysqlclient
pycountry
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.1
reportlab==3.5.56
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
urllib3==1.25.10
virtualenv==20.0.23
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
